refactor: extract shared utilities and DRY up wiki-server routes/tests
- Extract parseJsonBody, validationError, invalidJsonError, notFoundError
  to shared routes/utils.ts (was duplicated in 3 route files)
- Extract extractColumns, createQueryResult, createBaseMockSql, mockDbModule,
  postJson to shared __tests__/test-utils.ts (was duplicated in 3 test files)
- Add formatIdResponse/formatIdSummary helpers in ids.ts to eliminate
  5x repeated `E${row.numericId}` response object construction
- Update all route files to use shared helpers consistently

Net reduction: ~500 lines removed across routes and tests.
